+++ packages/content/src/runtime/app/components/ContentRendererInline.vue
@@ -5,7 +5,7 @@ import type { MarkdownRoot } from '../../../types/content'
 import { useRuntimeConfig, useState } from '#imports'
 import MarkdownRenderer from './internal/MarkdownRenderer.js'
 import { useUnwrap } from '../composables/useUnwrap'
-import { resolveMarkdownPlugins, resolveMarkdownRendererComponents } from '../../markdown/plugins'
+import { resolveMarkdownPlugins, resolveMarkdownRendererComponents, resolveMarkdownRendererFallbackComponents } from '../../markdown/plugins'
 import { toMarkdownRoot } from '../../../core/markdown/tree'
 import { parseComark } from '../../../core/markdown/parse-comark'
 import { loadContentComponentEntries } from '../../../integrations/vue/content-components'
@@ -107,6 +107,7 @@ watch(
     :tag="tag"
     :prose="prose"
     :components="resolvedComponents"
+    :fallback-components="resolveMarkdownRendererFallbackComponents()"
     v-bind="rendererAttrs"
   />
 </template>